general: add extra tests for new style type annotations
also update readme
